refactor: reduce component nesting and implement path resolver function
