+++ package.json
@@ -5,7 +5,7 @@
 	"main": "index.js",
 	"type": "module",
 	"scripts": {
-		"start": "node ./src/index.js -- --username=SomeUserName"
+		"start": "node ./src/index.js --username=SomeUserName"
 	},
 	"keywords": [],
 	"author": "",
